# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CampusOS Backend",
    "description": "\n    CampusOS Backend API \u2014 Trusted Operating System for African Universities.\n    Powered by Quai Network & Blip Pay.\n    \n    ### Flagship Module: Verified Student Identity (Milestone 2 & Milestone 3)\n    - Multi-document upload to Cloudinary (Student ID, Admission Letter)\n    - Institutional email (.edu.ng) validation\n    - Administrative verification queue & review workflow\n    - SHA-256 cryptographic credential hashing\n    - Quai Network on-chain credential registration (`StudentIdentity` smart contract)\n    - Permanent cryptographically signed Campus Identity QR\n    - Strict OWASP security headers, rate limiting, and magic-bytes file validation\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/payments/records/order/{order_id}": {
      "get": {
        "tags": [
          "Payments (Blip Pay & Escrow)"
        ],
        "summary": "Get audited Blip Pay payment records for an order",
        "description": "Returns chronological audit trail of all Blip Pay payment intents and webhooks associated with an order ID.",
        "operationId": "get_payment_records_by_order_api_v1_payments_records_order__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlipPaymentRecordResponse"
                  },
                  "title": "Response Get Payment Records By Order Api V1 Payments Records Order  Order Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/orders/": {
      "post": {
        "tags": [
          "Orders & Quai Escrow"
        ],
        "summary": "Create a new marketplace order",
        "description": "Creates a new order with line items in 'initiated' status.",
        "operationId": "create_order_api_v1_orders__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders": {
      "post": {
        "tags": [
          "Orders & Quai Escrow"
        ],
        "summary": "Create a new marketplace order",
        "description": "Creates a new order with line items in 'initiated' status.",
        "operationId": "create_order_api_v1_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/history": {
      "get": {
        "tags": [
          "Orders & Quai Escrow"
        ],
        "summary": "Get paginated order history for user",
        "description": "Returns chronological order history for a student as buyer, seller, or both.",
        "operationId": "get_order_history_api_v1_orders_history_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "UUID of the student",
              "title": "User Id"
            },
            "description": "UUID of the student"
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by role ('buyer', 'seller', 'all')",
              "default": "all",
              "title": "Role"
            },
            "description": "Filter by role ('buyer', 'seller', 'all')"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderResponse"
                  },
                  "title": "Response Get Order History Api V1 Orders History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}": {
      "get": {
        "tags": [
          "Orders & Quai Escrow"
        ],
        "summary": "Get order by ID",
        "description": "Returns order details, escrow status, and Quai Network escrow transaction hash.",
        "operationId": "get_order_by_id_api_v1_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/buyer/{buyer_id}": {
      "get": {
        "tags": [
          "Orders & Quai Escrow"
        ],
        "summary": "Get paginated orders for buyer",
        "description": "Returns chronological list of orders placed by the student.",
        "operationId": "get_orders_by_buyer_api_v1_orders_buyer__buyer_id__get",
        "parameters": [
          {
            "name": "buyer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Buyer Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderResponse"
                  },
                  "title": "Response Get Orders By Buyer Api V1 Orders Buyer  Buyer Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/seller/{seller_id}": {
      "get": {
        "tags": [
          "Orders & Quai Escrow"
        ],
        "summary": "Get paginated orders for seller",
        "description": "Returns chronological list of orders received by the seller.",
        "operationId": "get_orders_by_seller_api_v1_orders_seller__seller_id__get",
        "parameters": [
          {
            "name": "seller_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Seller Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderResponse"
                  },
                  "title": "Response Get Orders By Seller Api V1 Orders Seller  Seller Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}/cancel": {
      "post": {
        "tags": [
          "Orders & Quai Escrow"
        ],
        "summary": "Cancel marketplace order",
        "description": "Cancels an order that is initiated or escrow_locked.",
        "operationId": "cancel_order_api_v1_orders__order_id__cancel_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "UUID of the buyer, seller, or admin",
              "title": "Actor Id"
            },
            "description": "UUID of the buyer, seller, or admin"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}/confirm-shipment": {
      "post": {
        "tags": [
          "Orders & Quai Escrow"
        ],
        "summary": "Confirm shipment of order item",
        "description": "Transitions order from 'escrow_locked' or 'escrow_funded' to 'shipped_pending_delivery'.",
        "operationId": "confirm_shipment_api_v1_orders__order_id__confirm_shipment_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "UUID of the seller",
              "title": "Actor Id"
            },
            "description": "UUID of the seller"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/orders/{order_id}/release-escrow": {
      "post": {
        "tags": [
          "Orders & Quai Escrow"
        ],
        "summary": "Release Quai smart contract escrow",
        "description": "Releases escrow funds on Quai Network, marks order as completed, and awards +5 Trust Score to both Buyer and Seller.",
        "operationId": "release_escrow_api_v1_orders__order_id__release_escrow_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "UUID of the buyer or administrator",
              "title": "Actor Id"
            },
            "description": "UUID of the buyer or administrator"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}/dispute": {
      "post": {
        "tags": [
          "Orders & Quai Escrow"
        ],
        "summary": "Dispute marketplace order",
        "description": "Opens formal dispute on escrow order for administrative governance review.",
        "operationId": "dispute_order_api_v1_orders__order_id__dispute_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "UUID of the buyer or seller",
              "title": "Actor Id"
            },
            "description": "UUID of the buyer or seller"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDisputeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlipPaymentRecordResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "order_id": {
            "type": "string",
            "title": "Order Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "payment_reference": {
            "type": "string",
            "title": "Payment Reference"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "blip_pay"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_id",
          "user_id",
          "payment_reference",
          "amount",
          "currency",
          "status",
          "created_at"
        ],
        "title": "BlipPaymentRecordResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderCreateRequest": {
        "properties": {
          "buyer_id": {
            "type": "string",
            "title": "Buyer Id",
            "description": "UUID of the buyer"
          },
          "listing_id": {
            "type": "string",
            "title": "Listing Id",
            "description": "UUID of the marketplace listing"
          },
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "description": "Agreed purchase amount"
          },
          "quantity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Quantity",
            "description": "Quantity of items purchased",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "buyer_id",
          "listing_id",
          "amount"
        ],
        "title": "OrderCreateRequest"
      },
      "OrderDisputeRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 5,
            "title": "Reason",
            "description": "Explanation for disputing order"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "OrderDisputeRequest"
      },
      "OrderItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "order_id": {
            "type": "string",
            "title": "Order Id"
          },
          "listing_id": {
            "type": "string",
            "title": "Listing Id"
          },
          "seller_id": {
            "type": "string",
            "title": "Seller Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "price_per_unit": {
            "type": "number",
            "title": "Price Per Unit"
          },
          "subtotal": {
            "type": "number",
            "title": "Subtotal"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_id",
          "listing_id",
          "seller_id",
          "quantity",
          "price_per_unit",
          "subtotal",
          "created_at"
        ],
        "title": "OrderItemResponse"
      },
      "OrderResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "buyer_id": {
            "type": "string",
            "title": "Buyer Id"
          },
          "listing_id": {
            "type": "string",
            "title": "Listing Id"
          },
          "seller_id": {
            "type": "string",
            "title": "Seller Id"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "payment_reference": {
            "type": "string",
            "title": "Payment Reference"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "escrow_tx_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escrow Tx Hash"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "listing_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Title"
          },
          "buyer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buyer Name"
          },
          "seller_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seller Name"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/OrderItemResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "id",
          "buyer_id",
          "listing_id",
          "seller_id",
          "amount",
          "payment_reference",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "OrderResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}